Add the variant price to the total in the top bar.
