+++ js/views/modals/purchase/Receipt.js
@@ -1,4 +1,3 @@
-import _ from 'underscore';
 import app from '../../../app';
 import loadTemplate from '../../../utils/loadTemplate';
 import BaseView from '../../baseVw';
@@ -17,6 +16,10 @@ export default class extends BaseView {
 
     if (!options.listing || !(options.listing instanceof Listing)) {
       throw new Error('Please provide a listing model');
+    }
+
+    if (!options.prices) {
+      throw new Error('Please provide the prices array');
     }
 
     this.listenTo(this.model.get('items').at(0), 'change', () => this.render());
@@ -39,43 +42,13 @@ export default class extends BaseView {
     this._coupons = filteredCoupons.map(coupon => coupon.toJSON());
   }
 
-  get prices() {
-    // create an array of price objects that matches the items in the order
-    const prices = [];
-    this.model.get('items').forEach(item => {
-      const priceObj = {};
-      const shipping = item.get('shipping');
-      const sName = shipping.get('name');
-      const sService = shipping.get('service');
-      const sOpt = this.options.listing.get('shippingOptions').findWhere({ name: sName });
-      const sOptService = sOpt ? sOpt.get('services').findWhere({ name: sService }) : '';
-      // determine which skus match the chosen options
-      const variantCombo = [];
-      item.get('options').forEach((option, i) => {
-        const variants = this.options.listing.get('item').get('options').at(i)
-          .get('variants')
-          .toJSON();
-        const variantIndex = variants.findIndex(variant => variant.name === option.get('value'));
-        variantCombo.push(variantIndex);
-      });
-      const sku = this.options.listing.get('item').get('skus').find(v =>
-        _.isEqual(v.get('variantCombo'), variantCombo));
-
-      priceObj.price = this.options.listing.get('item').get('price');
-      priceObj.sPrice = sOptService ? sOptService.get('price') : 0;
-      priceObj.vPrice = sku ? sku.get('surcharge') : 0;
-      prices.push(priceObj);
-    });
-    return prices;
-  }
-
   render() {
     loadTemplate('modals/purchase/receipt.html', t => {
       this.$el.html(t({
         listing: this.options.listing.toJSON(),
         coupons: this.coupons,
         displayCurrency: app.settings.get('localCurrency'),
-        prices: this.prices,
+        prices: this.options.prices,
         ...this.model.toJSON(),
       }));
     });
